% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpopt.R
\name{cpopt}
\alias{cpopt}
\title{Find the optimal complexity parameter value}
\usage{
cpopt(df, conserve = FALSE)
}
\arguments{
\item{df}{A dataframe with columns of Start.Pos and log2r}

\item{conserve}{The default is FALSE.  When TRUE the optimal complexity parameter has a minimum around .01 to reduce the chance of overfitting.
When FALSE the optimal complexity has a minimum around .001.}
}
\value{
The optimal complexity parameter in numeric format
}
\description{
The function finds the complexity parameter that is most optimal for reducing cross validation error.
}
\examples{
 my <- cpopt(df)
 print(my)
}
\author{
Annika Cleven
}
