% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterseg.chr.R
\name{iterseg.chr}
\alias{iterseg.chr}
\title{Segment and Visualize a Single Chromosome}
\usage{
iterseg.chr(df, chromid, cpvalue = NA)
}
\arguments{
\item{df}{A dataframe with columns of Start.Pos, log2r, and Chr columns.  The Chr column should have format like "chr1", "chr21", "chrY".}

\item{chromid}{The chromosome of interest in the form "chr3" or "chr21"}

\item{cpvalue}{Specify a constant cp value for the regression tree to use instead of the optimal cp value}
}
\value{
A list containing a dataframe with all of the segmentation data (segments),
and a dataframe with the predictions from the regression tree (regtreepred),
a plot of the chromosome data and final predictions (chrplot), a list 5 plots of each step in iteration (plots)
}
\description{
Creates a plot of the chromosome of interest segmented, a dataframe including all the segments,
and a dataframe including the predictions from the regression tree using the optimal cp value for the chromosome,
and iterative regression trees.
}
\examples{
example <- iterseg.chr(dataframe, "example.png")
example$regtreepred
example$segments
example$plots$iter2
}
\author{
Annika Cleven
}
