% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterseg.genome.R
\name{iterseg.genome}
\alias{iterseg.genome}
\title{Segment and Visualize a Whole Genome}
\usage{
iterseg.genome(
  df,
  png_filename,
  cpvalue = NA,
  upper.y.lim = 5,
  lower.y.lim = 5
)
}
\arguments{
\item{df}{A dataframe with columns of Start.Pos, log2r, and Chr columns.  The Chr column should have format like "chr1", "chr21", "chrY".}

\item{png_filename}{A .png filename}

\item{cpvalue}{Specified constant cp value for the regression tree to use instead of the optimal cp value}

\item{upper.y.lim}{The upper limit for the y value of the plot}

\item{lower.y.lim}{The lower limit for the y value of the plot}
}
\value{
A .png file with the whole genome plot and segmentation data,
a list containing a dataframe with all of the segmentation data (segments),
and a dataframe with the predictions from the regression tree (regtreepreds)
}
\description{
Creates a plot of the Whole Genome segmented, a dataframe including all the segments,
and a dataframe including the predictions from the regression tree using the optimal cp value for each chromosome
and iterative regression trees.
}
\examples{
example <- iterseg.genome(s131L001, "example.png")
example$regtreepred
example$segments
}
\author{
Annika Cleven
}
