% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.genome.R
\name{seg.genome}
\alias{seg.genome}
\title{Segment and Visualize a Whole Genome}
\usage{
seg.genome(df, png_filename, upper.y.lim = 5, lower.y.lim = -5, cpvalue = NA)
}
\arguments{
\item{df}{A dataframe with columns of Start.Pos, log2r, and Chr columns.  The Chr column should have format like "chr1", "chr21", "chrY".}

\item{png_filename}{A .png filename}

\item{upper.y.lim}{The upper limit for the y value of the plot}

\item{lower.y.lim}{The lower limit for the y value of the plot}

\item{cpvalue}{Specified constant cp value for the regression tree to use instead of the optimal cp value}
}
\value{
A .png file with the whole genome plot and segmentation,
a list containing a dataframe with all of the segmentation data(segments),
and a dataframe with the predictions from the regression tree(regtreepred)
}
\description{
Creates a plot of the Whole Genome segmented, a dataframe including all the segments,
and a dataframe including the predictions from the regression tree using the optimal cp value for each chromosome
}
\examples{
example <- seg.genome(datafr127, png_filename = "datafr.png",upper.y.lim = 5, lower.y.lim = -5)
example$regtreepred
example$segments
}
\author{
Annika Cleven
}
