% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeGenome.Plot.R
\name{WholeGenome.Plot}
\alias{WholeGenome.Plot}
\title{Plotting Whole Genome}
\usage{
WholeGenome.Plot(
  png_filename,
  chr,
  x,
  s = NA,
  segmentnumber = NA,
  sample.name = "sample",
  ylab = "Allele Imbalance",
  BAF = F,
  up.y = 3,
  lo.y = -3,
  plot.log = "",
  odd.col = "magenta",
  even.col = "dodgerblue",
  x.col = "blueviolet",
  y.col = "darkmagenta",
  plot.cex = 3,
  plot.pch = ".",
  plot.bg.color = "white",
  is.chr.labeled = TRUE,
  is.txt.output = TRUE,
  output.digit = 3,
  odd.y.pos = -10,
  even.y.pos = -38,
  label.cex = 1,
  med.line.col = "red"
)
}
\arguments{
\item{png_filename}{A .png filename that the plot will be exported to}

\item{chr}{The chromosome the datapoint is a part of}

\item{x}{The data point value that is being plotted}

\item{s}{The smoothing value}

\item{segmentnumber}{The number of segments in the tree, the value here will be printed under the title of the plot.}

\item{sample.name}{The name of the sample}

\item{ylab}{The y axis title}

\item{up.y}{The y axis upper limit}

\item{lo.y}{The y axis lower limit}

\item{odd.col}{The color of the odd numbered chromosomes}

\item{even.col}{The color of the even numbered chromosomes}

\item{x.col}{The color of the x chromsome}

\item{y.col}{The color of the y chromosome}

\item{plot.bg.color}{The color of the plot background}

\item{med.line.col}{The color of the median line}
}
\value{
A .png file with the whole genome data
}
\description{
This plots whole genome data with options to smooth, color, and other customizations
}
\examples{
  WholeGenome.Plot("test.png", chr = data$Chr ,x = data$log2r, up.y=5, lo.y = -5)
}
\author{
Annika Cleven
}
